falta calculo de determinantes, inversão da matriz e calculo dos cofatores

+++ inc/minirt.h
@@ -44,9 +44,9 @@ typedef struct s_color
 
 typedef struct s_matrix
 {
-	int				**matrix;
-	int				collum;
-	int				row;
+	double				**matrix;
+	double				collum;
+	double				row;
 }					t_matrix;
